Create h5 directory in include and move relevant files
